backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.7
requests==2.31.0
python-dotenv==1.0.1
django-autoslug==1.9.9